requirements.txt
discord.py>=1.7.3,<2
tomlkit>=0.8.0,<1
beautifulsoup4>=4.10.0,<5
requests>=2.27.1,<3
